Move Design Selector to separate file
